Initial sound emulation for WPF and XNA+Windows platforms via COM interop to DirectSound. Added rudimentary UI to Silverlight and WPF platforms to enable disk selection at runtime.
--HG--
extra : convert_revision : svn%3Affd33b8c-2492-42e0-bdc5-587b920b7d6d/trunk%4024459

+++ Virtu/CpuData.cs
@@ -7,7 +7,7 @@ namespace Jellyfish.Virtu
         private const int CyclesPerUpdate = 17030;
         private const int CyclesPerVSync = 17030;
         private const int CyclesPerSecond = 1022730;
-        private static readonly long TicksPerVSync = TimeSpan.FromSeconds((double)CyclesPerVSync / (double)CyclesPerSecond).Ticks;
+        private static readonly long TicksPerVSync = TimeSpan.FromSeconds((double)CyclesPerVSync / CyclesPerSecond).Ticks;
 
         private const int OpcodeCount = 256;
